Rename & move Home.jsx, refactor ternary for rendering create new list button
