+++ src/App.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import { AddItemView } from './pages/AddItemView';
 import { ListView } from './pages/ListView';
-import Home from './components/Home';
+import Home from './pages/Home';
 import { RouteLink } from './components/RouteLink';
 import './App.css';
